requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-groq
chromadb
pypdf
sentence-transformers
